refactor: 뒤로 가는 버튼의 재활용성을 위한 컴포넌트 생성
